+++ sysml2-mode.el
@@ -2,7 +2,7 @@
 
 ;; Copyright (C) 2026 sysml2-mode contributors
 ;; Author: sysml2-mode contributors
-;; Version: 0.5.4
+;; Version: 0.5.5
 ;; Package-Requires: ((emacs "29.1"))
 ;; Keywords: languages, systems-engineering, sysml
 ;; URL: https://github.com/jackhale98/sysml2-mode
@@ -98,7 +98,7 @@
 
 ;; --- Version ---
 
-(defconst sysml2-mode-version "0.5.4"
+(defconst sysml2-mode-version "0.5.5"
   "Version of sysml2-mode.")
 
 (defun sysml2-version ()